+++ src/main/java/dev/dfonline/flint/Flint.java
@@ -1,7 +1,5 @@
 package dev.dfonline.flint;
 
-import dev.dfonline.flint.actiondump.ActionDump;
-import dev.dfonline.flint.actiondump.ActionType;
 import dev.dfonline.flint.feature.core.FeatureManager;
 import dev.dfonline.flint.feature.core.FeatureTrait;
 import dev.dfonline.flint.feature.core.FeatureTraitType;
@@ -33,8 +31,6 @@ import net.fabricmc.fabric.api.client.rendering.v1.world.WorldRenderEvents;
 import net.kyori.adventure.platform.modcommon.MinecraftAudiences;
 import net.kyori.adventure.platform.modcommon.MinecraftClientAudiences;
 import net.minecraft.client.MinecraftClient;
-
-import java.util.Arrays;
 
 public class Flint implements ClientModInitializer {
