+++ src/main/java/dev/dfonline/flint/actiondump/codeblocks/BlockTagType.java
@@ -1,4 +1,6 @@
-package dev.dfonline.flint.actiondump;
+package dev.dfonline.flint.actiondump.codeblocks;
+
+import dev.dfonline.flint.actiondump.Icon;
 
 public record BlockTagType(
         String name,
