+++ src/main/java/dev/dfonline/flint/actiondump/codeblocks/CodeBlockType.java
@@ -1,4 +1,6 @@
-package dev.dfonline.flint.actiondump;
+package dev.dfonline.flint.actiondump.codeblocks;
+
+import dev.dfonline.flint.actiondump.Icon;
 
 public record CodeBlockType(
         String name,
